+++ PAF/snippets/core/TEMPLATE_Get-Example.ps1
@@ -1,0 +1,28 @@
+<#
+.SYNOPSIS
+
+shor description. it is shown in menu.
+
+.DESCRIPTION
+
+long description
+#>
+
+
+function function_name_as_verb-noun {
+<#
+:CATEGORY
+<put category here>
+
+:NAME
+<function_name_as_verb-noun, i.e. Get-Example>
+#>
+    [CmdletBinding()]
+    param (    )
+
+    # Here goes the code for the function snippet
+
+}
+
+# The snippet must have code to run the function(s) when invoked by PAF.
+Get-Example